+++ tools/express-server.js
@@ -1,27 +1,38 @@
-const express = require('express')
-const esm = require('esm')
-const { getArg } = esm(module)(require.resolve('./utils'))
-const app = express()
+const { paramCase } = require('change-case');
+const express = require('express');
+const cors = require('cors');
+const app = express();
 
-const dist = getDist()
-const host = getHost()
-const port = getPort()
+app.use(cors());
+app.disable('x-powered-by');
 
-app.use(express.static(dist))
-app.listen(port, host, function () {
-  const { address, port } = this.address()
+const dist = getDist();
+
+app.use(express.static(dist));
+app.listen(getPort(), getHost(), function () {
+  const { address, port } = this.address();
   // eslint-disable-next-line no-console
-  console.log(`server listening on http://${address}:${port} ; dist: ${dist}`)
-})
+  console.log(`server listening on http://${address}:${port} ; dist: ${dist}`);
+});
 
 function getDist () {
-  return process.env.npm_config_dist || getArg('dist') || process.env.DIST || 'dist'
+  return process.env.npm_config_dist || getArg('dist') || process.env.DIST || 'dist';
 }
 
 function getHost () {
-  return process.env.npm_config_host || getArg('host') || process.env.HOST || 'localhost'
+  return process.env.npm_config_host || getArg('host') || process.env.HOST || 'localhost';
 }
 
 function getPort () {
-  return process.env.npm_config_port || getArg('port') || process.env.PORT || 3000
+  return process.env.npm_config_port || getArg('port') || process.env.PORT || 3000;
+}
+
+function getArg (name) {
+  name = paramCase(name);
+  const args = process.argv.slice(2);
+  let value = !!args[args.indexOf(`--${name}`)];
+  if (args.includes(`--${name}`) && args[args.indexOf(`--${name}`) + 1]) {
+    value = args[args.indexOf(`--${name}`) + 1];
+  }
+  return value;
 }